pyproject.toml
[tool.poetry]
name = "revel"
version = "0.1.0"
description = "A Remote Development revel tool"
authors = ["Gonzalo Peci <pecigonzalo@users.noreply.github.com>"]
readme = 'README.md'

[tool.poetry.scripts]
revel = "revel.cli:app"

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "^1.20.26"
typer = { extras = ["all"], version = "^0.4.0" }
PyYAML = "^6.0"
halo = "^0.0.31"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.12b0"
isort = "^5.10.1"
mypy = "^0.930"
flake8 = "^4.0.1"
flake8-bugbear = "^21.11.29"
pytest-mock = "^3.6.1"
types-setuptools = "^57.4.7"
types-PyYAML = "^6.0.3"
types-boto3 = "^1.0.1"
boto3-stubs = { extras = ["ec2"], version = "^1.20.31" }
moto = { extras = ["ec2", "iam"], version = "^2.3.1" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"

[tool.mypy]
follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
